requirements.txt
# Core web framework and API
fastapi==0.95.1
uvicorn==0.22.0

# Task queue and background processing
celery==5.2.7
redis==4.5.5

# Database and ORM
sqlalchemy==2.0.12
alembic==1.10.4
psycopg2-binary==2.9.6

# Configuration and environment
pydantic==1.10.7
python-dotenv==1.0.0

# Logging
structlog==23.1.0

# CLI and utilities
typer==0.9.0

# Cloud storage and file handling
boto3==1.28.0
aiofiles==23.1.0
python-multipart==0.0.6

# OCR and document processing
pillow==9.5.0
pytesseract==0.3.10
paddleocr==2.7.3
pdf2image==1.16.3
amazon-textract-textractor==1.0.56

# Google APIs
google-auth==2.16.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
google-cloud-documentai==2.16.0

# Azure AI
azure-ai-documentintelligence==1.0.0b1

# Security and cryptography
cryptography==40.0.2

# HTTP client for external APIs
httpx==0.24.1

# Additional utilities
click==8.1.3
rich==13.4.2

# File type detection and encoding
python-magic==0.4.27
chardet==5.1.0
